restore preprocessors as separate from utils
